Bug fix
- dropped reference sign in some calls by reference.
- some indexing minor bugs.
- use history index iterators instead of normal ones.
- small SAValue for path index (it was configured such that it could
  contain sequences of length at most 2^16 = 65536 which is too small
  for maximal paths).
- fix `kmer_exact_matches` function for TopDown fine index iterators.
- fix wrong order of parameters in calling `seqan::readRecords` method.
- and some other minor fixes.

+++ src/index.h
@@ -40,7 +40,7 @@ namespace seqan {
   template < >
   struct SAValue< StringSet < Dna5QString > >
   {
-    typedef Pair<long unsigned int, uint16_t, Tag<Pack_> > Type;
+    typedef Pair<long unsigned int, long unsigned int, Tag<Pack_> > Type;
   };
 }  /* -----  end of namespace seqan  ----- */
